+++ microgw-20200810/ChangeLog.md
@@ -1,3 +1,6 @@
+2020-11-16 Version: 20201116
+- Add params for FindGateways.
+
 2020-11-06 Version: 1.0.0
 - Java SDK.